Correct `householder_product` docs.
When discussing https://github.com/pytorch/pytorch/pull/63880, we
realised that the docs of `householder_product` were not correct. This
PR fixes this.

The new docs are slightly more difficult, but hopefully correct. Note
that this is a LAPACK function in disguise, so it is expected the
specification to be more difficult than normal.

ghstack-source-id: e62da39a25e7afbb2fe5113211aac81809c7c1ba
Pull Request resolved: https://github.com/pytorch/pytorch/pull/68335