Remove CUDA Foreach... files dependency on function operators
This adds a new macros `TORCH_ASSERT_ONLY_METHOD_OPERATORS` which
allows `Tensor.h` to be included, but not other operators. So, a file
that defines this only needs to be re-compiled when a method operator
is changed.

I also apply it to the `Foreach...cu` files which would be hard to
rewrite using `TensorBase` since they return `std::vector<Tensor>`.

ghstack-source-id: 05e5e68cf8056a00d6e41b76c57b06c50c18a47e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/68462