dbr quant overhead[3/x]: speed up AutoQuantizationState.mark_cur_op_complete
Summary:

`torch.nn.Module` has overhead for getting and setting attributes because
it does various type checks on the attribute.

This PR explicitly gets and sets the right thing for this particular
function, avoding the type checks. Model level benchmarks are too noisy,
but according to function level profiling this reduces the time spent in
this function in a quantized model from 2.60% to 0.53%, on MobileNetV2 with
input size 1x3x224x224.

Test plan:

```
python test/test_quantization.py TestQuantizeDBR
```

ghstack-source-id: ee7681d190eb8aa20d81d8d3542fc9acca59e6ab
Pull Request resolved: https://github.com/pytorch/pytorch/pull/68350