dbr quant overhead[5/x]: remove unnecessary asserts
Summary:

Removes asserts which are duplicate (the same condition is checked
when calculating the hook type, so there is no need to check it again).
For the assert in `validate_is_at_last_seen_idx`, rewrites it to
raise an Error instead to ensure it does not get stripped in
production environments.

Test plan:

```
python test/test_quantization.py TestQuantizeDBR
```

ghstack-source-id: 037c2a7b1c2fe43de998f43df9433128b5b77518
Pull Request resolved: https://github.com/pytorch/pytorch/pull/68370